+++ database/migrations/2024_08_07_110557_create_orders_table.php
@@ -17,7 +17,7 @@ return new class extends Migration {
             $table->double('amount')->default(0.00)->comment("total order amount");
             $table->text('info')->nullable()->comment("for dynamic information store");
             $table->boolean('payment_status')->default(0)->comment("0=>incomplete,1=>complete");
-            $table->tinyInteger('status')->default(0)->comment("0=>initiate,1=>complete,2=>refund");
+            $table->integer('status')->default(0)->comment("0=>initiate,1=>complete,2=>refund");
             $table->enum('order_for', ['topup', 'card']);
             $table->string('utr')->nullable();
             $table->timestamps();
