+++ database/migrations/2024_08_22_182852_create_reviews_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->id();
             $table->morphs('reviewable');
             $table->foreignId('user_id')->index();
-            $table->tinyInteger('rating');
+            $table->integer('rating');
             $table->text('comment');
             $table->boolean('status')->default(1)->comment("0=>inactive,1=>active");
             $table->timestamps();
